요청값 주문 가격 타입을 long으로 변경, 주문 정보 저장 시 배송 정보도 같이 저장하도록 변경

+++ src/main/java/shop/goodspia/goods/service/IamportService.java
@@ -17,6 +17,7 @@ import shop.goodspia.goods.exception.PaymentValidationFailureException;
 import shop.goodspia.goods.repository.OrderQueryRepository;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 
 @Slf4j
 @Service
@@ -39,7 +40,7 @@ public class IamportService implements PaymentAgentService {
 
         //아임포트 API로 주문번호과 결제금액 사전등록
         Prepare prepare = iamportClient.postPrepare(
-                new PrepareData(prepareDto.getMerchantUid(), prepareDto.getAmount())).getResponse();
+                new PrepareData(prepareDto.getMerchantUid(), BigDecimal.valueOf(prepareDto.getAmount()))).getResponse();
 
         return parsingToPaymentPrepareResponseDto(prepare);
     }
